Simple fix for the incorrect sizing problem.
This actually can result in bad RPS sizes in register 74.w (RPS)
for slots where
there is no DRAM installed. This really does not matter since the actual
access to DRAM is controlled elsewhere. If there is not DRAM in a slot,
then that piece of the RPS register will never be used.